+++ feedcore/CMakeLists.txt
@@ -10,7 +10,6 @@ set(feedcore_HEADERS
     context.h
     scheduler.h
     aggregatefeed.h
-    allitemsfeed.h
     categoryfeed.h
     factory.h
     provisionalfeed.h
@@ -33,7 +32,6 @@ set(feedcore_SRCS
     context.cpp
     scheduler.cpp
     aggregatefeed.cpp
-    allitemsfeed.cpp
     categoryfeed.cpp
     provisionalfeed.cpp
     networkaccessmanager.cpp
