+++ src/application.cpp
@@ -215,8 +215,8 @@ void Application::onLastWindowClosed()
         quit();
         return;
     }
-    unloadEngine();
     startBackgroundNotifier();
+    unloadEngine();
 #endif
 }
