+++ src/notificationcontroller.cpp
@@ -4,7 +4,6 @@
  */
 
 #include "notificationcontroller.h"
-#include "allitemsfeed.h"
 #include "article.h"
 #include "cmake-config.h"
 #include "context.h"
@@ -19,7 +18,7 @@
 using namespace FeedCore;
 
 struct NotificationController::PrivData {
-    FeedCore::Feed *feed{nullptr};
+    QSharedPointer<Feed> feed{nullptr};
     int counter{0};
 
 #ifdef HAVE_SYSTEM_TRAY
@@ -32,9 +31,9 @@ NotificationController::NotificationController(FeedCore::Context *context, QObje
     : QObject(parent)
     , d{std::make_unique<PrivData>()}
 {
-    d->feed = new AllItemsFeed(context, "", this);
-    QObject::connect(d->feed, &Feed::statusChanged, this, &NotificationController::onStatusChanged);
-    QObject::connect(d->feed, &Feed::articleAdded, this, &NotificationController::onArticleAdded);
+    d->feed = context->allItemsFeed();
+    QObject::connect(d->feed.get(), &Feed::statusChanged, this, &NotificationController::onStatusChanged);
+    QObject::connect(d->feed.get(), &Feed::articleAdded, this, &NotificationController::onArticleAdded);
 
 #ifdef HAVE_SYSTEM_TRAY
     d->trayMenu = std::make_unique<QMenu>();
